fix another assertion

+++ migration-engine/migration-engine-tests/tests/migrations/squashing_tests.rs
@@ -64,15 +64,16 @@ async fn squashing_whole_migration_history_works(api: &TestApi) -> TestResult {
 
     for entry in std::fs::read_dir(directory.path())? {
         let entry = entry?;
+        if entry.metadata()?.is_dir() {
+            assert!(entry.metadata()?.is_dir());
 
-        assert!(entry.metadata()?.is_dir());
+            let file_path = entry.path().join("migration.sql");
+            let contents = std::fs::read_to_string(file_path)?;
 
-        let file_path = entry.path().join("migration.sql");
-        let contents = std::fs::read_to_string(file_path)?;
+            squashed_migrations.push((entry.file_name().into_string().unwrap(), contents));
 
-        squashed_migrations.push((entry.file_name().into_string().unwrap(), contents));
-
-        std::fs::remove_dir_all(entry.path())?;
+            std::fs::remove_dir_all(entry.path())?;
+        }
     }
 
     squashed_migrations.sort_by(|left, right| left.0.cmp(&right.0));
